Use TopDownMutator instead of LoadHook
Soong's "Load Hooks" allow creation of further modules, but because they
are the first thing to run, they do not permit any other processing.

In order to support dependencies between modules other than what Soong
provides, we will need to to run mutators before creating the Soong
modules. Make this possible by changing the load hooks into a normal
top-down mutator pass.

This also moves the `Main()` function into `standalone.go` - this is
required so that we can have a separate implementation of
`featureApplierMutator` and `templateApplierMutator` for each backend.

Rename `ctx` to `mctx` in the former load hooks for consistency with
other mutators' context variables.
